backend/requirements.txt
qrcode
pillow
Flask
flask-cors
reportlab
web3
python-dotenv
gunicorn
